refactor main.py into separate model, api, and client files
