+++ backend/app/clients/openai.py
@@ -1,0 +1,15 @@
+from openai import OpenAI
+
+client = OpenAI()
+
+
+async def fetch_definition(lang: str, word: str):
+    prompt = (
+        f"For the word '{word}' in '{lang}', "
+        "please provide a definition and an example sentence."
+    )
+    response = await client.responses.create(
+        model="o4-mini-2025-04-16",
+        input=prompt,
+    )
+    return response
