+++ backend/app/models/word_data.py
@@ -1,0 +1,9 @@
+# All models related to a word/vocab
+# Includes example sentences, their translations, etc.
+
+from pydantic import BaseModel
+
+
+class Word(BaseModel):
+    lang: str
+    word: str
